Change props to be injected into Machine#start()
This allows you to keep the same machine, but re-run it with different
props. It's also useful for the immediate intended use case of games:
scene/actor based models often differentiate between construction time,
when various game subsytems aren't yet initialized, and some kind of
"initialize" time, where you can actually initialize objects in the
systems. We want to be able to pass the initialized objects in as props,
but to do so, we'd need to re-instantiate the machine every time, rather
than defining the machine at construction time and initializing it at
init-time.