Refactor props to make start/stop less special
It was neat that props were passed into start and stop, but this was
annoying for any other methods you might define on state classes: since
you only had access to the props in `start` and `stop`, you'd need to
store the variable somewhere if you wanted to access it elsewhere. Since
the value wasn't known at constructor time, though, that meant you never
could have guaranteed typesafe access: either you'd need if statements
to guard against it being non-nullish (in which case, if somehow it was
nullish, your code wouldn't run), or you'd need a bunch of type casts to
avoid the compiler yelling at you.

Instead, we now pass props into the constructor, since we now
instantiate the states on-demand rather than at machine construction
time. Then we store the props in a readonly variable on the state
instance. Since it's guaranteed to be exactly what you've defined as
`Props`, you can access it safely in `start`, `stop`, or any other
methods you define.