Optimize nested machines and make them simpler
Use a single `children` map, rather than trying to introspect on the
state class on every transition. This makes the feature essentially free
for people who don't use it (only a single runtime check, and since the
property is read-only it'll be very friendly to the branch predictor),
and much simpler for people who do use it (it's not possible to forget
to call `.addChild` on a field, since that method no longer exists, and
any children in the `.children` property will be recognized).